Add command and task to compile Inform 6 files, with a problem matcher
TypeScript dependencies has been added, and launch.json and tasks.json has been modified to take account of it. Also, new config files for TypeScript, ESLint and Webpack.

.gitignore has been added and .vscodeignore edited to take account of these changes.

Activation events have been added. A command and a task to compile has been added, as well as menus contributions to launch this action. New problem matcher for the new task.

Two new settings have been added: one for the Inform 6 compiler path, and one for the arguments that are passed to it.