Some fixes and changes in syntax highlighting
`move x to` and `Array x ->/-->/buffer/string/table` are now properly highlighted (the arrows are coloured and there have to be word boundaries at the edges.)

Binary and hex number greater than $FFFF are now highlighted.

Characters consisting of a Unicode escaped are now properly coloured as characters instead of dict words.

JavaScript string when executing JS commands with Vorple are not highlighted anymore. It caused problems when concatenating strings with `BuildCommand`.