Some more fixes and changes in syntax highlighting
`Zcharacter table` directive is now highlighted.

`Version` is not marked as a deprecated directive since it's also an action.

`replace` is no longer a keyword since the directive had precedence anyway.

Unicode @-escapes with a value too big (> 7 digits) are now marked as illegal, as well as single `@`s in strings, words and chars.

Custom opcodes (e.g. `@"S2:729"`) are now highlighted.