Add opening of story after compiling + refactor how tasks are managed
There's a new TaskManager class that takes care of tasks.

The cwd of the execution of Inform 6 tasks is now set to the location of the source (so that the compiled story is next to the source).

Stories can now be opened after compilation. There's a new configuration for that.